enable -Werror for MinGW CMake jobs

See merge request graphviz/graphviz!2294

+++ lib/sfio/sfwr.c
@@ -10,6 +10,7 @@
 
 #include	<sfio/sfhdr.h>
 #include	<stddef.h>
+#include	<stdint.h>
 
 /*	Write with discipline.
 **
@@ -33,7 +34,7 @@ static ssize_t sfoutput(Sfio_t * f, char *buf, size_t n)
 	    while ((ssize_t) n >= _Sfpage) {	/* see if a hole of 0's starts here */
 		sp = buf + 1;
 		if (buf[0] == 0 && buf[_Sfpage - 1] == 0) {	/* check byte at a time until int-aligned */
-		    while (((ulong) sp) % sizeof(int)) {
+		    while ((uintptr_t)sp % sizeof(int)) {
 			if (*sp != 0)
 			    goto chk_hole;
 			sp += 1;
